+++ src/shared/fileNames.ts
@@ -445,6 +445,7 @@ export default {
   "tsconfig.spec.json": "_f_tsconfig",
   "tsconfig.staging.json": "_f_tsconfig",
   "tsconfig.test.json": "_f_tsconfig",
+  "tsconfig.tsd.json": "_f_tsconfig",
   "tslint.json": "_f_tslint",
   "tslint.yaml": "_f_tslint",
   "tslint.yml": "_f_tslint",