+++ src/shared/fileExtensions.ts
@@ -494,4 +494,5 @@ export default {
   deflate: "_f_zip",
   brotli: "_f_brotli",
   kra: "_f_krita",
+  mgcb: "_f_mgcb",
 };