+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Add Atomizer icon
 - Add Esbuild icon
 - Add tox icon
+- Add Mix icons
 
 ## 1.6.0 - 2023-03-12